L2/RAG Systems Essentials/requirements.txt
# RAG Systems Essentials - Dependencies
# Portfolio-grade RAG system for research paper Q&A

# Core ML/NLP
sentence-transformers>=2.2.0     # Text embeddings
faiss-cpu>=1.7.4                 # Vector similarity search
torch>=2.0.0                     # Deep learning backend
transformers>=4.30.0             # Transformer models

# PDF Processing
pymupdf>=1.23.0                  # Fast PDF extraction (fitz)
pdfminer.six>=20221105           # Backup PDF extraction

# LLM Providers (NO OpenAI!)
google-generativeai>=0.3.0       # Gemini API
groq>=0.4.0                      # Groq API
requests>=2.28.0                 # For Ollama HTTP API

# Utilities
numpy>=1.24.0                    # Numerical operations
tqdm>=4.65.0                     # Progress bars

# Optional: Development
python-dotenv>=1.0.0             # Environment variables
